SolChain-Prototype/ai-ml-services/main.py
# Main entry point for AI/ML services
# TODO: FastAPI application for AI services

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
from typing import List, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="SolChain AI/ML Services",
    description="AI and Machine Learning services for energy forecasting, pricing, and optimization",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # TODO: Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# TODO: Import service modules
# from services.pricing_service import PricingService
# from services.forecasting_service import ForecastingService
# from services.anomaly_detection import AnomalyDetectionService
# from services.optimization_service import OptimizationService

@app.get("/")
async def root():
    return {"message": "SolChain AI/ML Services", "status": "operational"}

@app.get("/health")
async def health_check():
    # TODO: Implement health check
    return {"status": "healthy", "services": ["pricing", "forecasting", "anomaly_detection", "optimization"]}

# TODO: Pricing endpoints
@app.post("/api/pricing/dynamic")
async def calculate_dynamic_pricing(market_data: Dict[str, Any]):
    # TODO: Implement dynamic pricing calculation
    # TODO: Consider supply/demand, time of day, weather, grid prices
    pass

@app.post("/api/pricing/forecast")
async def forecast_energy_prices(historical_data: List[Dict[str, Any]]):
    # TODO: Implement price forecasting using ML models
    pass

# TODO: Energy forecasting endpoints
@app.post("/api/forecast/generation")
async def forecast_energy_generation(weather_data: Dict[str, Any], historical_data: List[Dict[str, Any]]):
    # TODO: Implement solar energy generation forecasting
    # TODO: Use weather data and historical patterns
    pass

@app.post("/api/forecast/demand")
async def forecast_energy_demand(usage_patterns: List[Dict[str, Any]]):
    # TODO: Implement energy demand forecasting
    # TODO: Consider historical usage, time patterns, seasonal effects
    pass

# TODO: Anomaly detection endpoints
@app.post("/api/anomaly/detect")
async def detect_anomalies(energy_data: List[Dict[str, Any]]):
    # TODO: Implement anomaly detection for energy patterns
    # TODO: Detect unusual consumption, generation, or trading patterns
    pass

@app.post("/api/anomaly/fraud")
async def detect_fraud(transaction_data: List[Dict[str, Any]]):
    # TODO: Implement fraud detection for energy transactions
    pass

# TODO: Optimization endpoints
@app.post("/api/optimize/distribution")
async def optimize_energy_distribution(grid_data: Dict[str, Any]):
    # TODO: Implement energy distribution optimization
    # TODO: Minimize losses, balance supply/demand
    pass

@app.post("/api/optimize/storage")
async def optimize_battery_storage(storage_data: Dict[str, Any]):
    # TODO: Implement battery storage optimization
    # TODO: Optimize charging/discharging cycles
    pass

# TODO: Analytics endpoints
@app.post("/api/analytics/efficiency")
async def calculate_efficiency_metrics(energy_data: List[Dict[str, Any]]):
    # TODO: Calculate system efficiency metrics
    pass

@app.post("/api/analytics/savings")
async def calculate_cost_savings(before_after_data: Dict[str, Any]):
    # TODO: Calculate cost savings from using SolChain platform
    pass

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", 8000)),
        reload=True
    )
