SolChain-Prototype/ai-ml-services/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.1
tensorflow==2.11.0
torch==1.13.1
web3==6.2.0
requests==2.28.2
python-dotenv==1.0.0
pydantic==1.10.6
sqlalchemy==2.0.7
asyncpg==0.27.0
redis==4.5.4
celery==5.2.7
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.0
